// proyecto-sibi/src/app/models/videojuego.ts
export interface videojuego {
  nombre: string;
  plataforma: string;
  lanzamiento: string;
  genero: string;
  empresa: string;
  puntuacionCritica: string;
  puntuacionUsuario: string;
  pegi: string;
  like: string;
  dislike: string;
  favorito: string;
  ranking: number;
  liked: boolean;
  disliked: boolean;
  fav: boolean;
}